data: Start split of registry with reducer
This splits out the implementation of `registerReducer` into a separate
file, with the intention of continuing the rest of the store
implementation. This will make room for other possible implementations
with a common interface.
